refactor and  add generic trainer class
