+++ habitat_baselines/__init__.py
@@ -3,3 +3,10 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
+
+# TODO include useful/reusable classes/functions
+
+# import for proper registry
+from habitat_baselines.rl.ppo.ppo_trainer import PPOTrainer, RolloutStorage
+
+__all__ = ["PPOTrainer", "RolloutStorage"]
