+++ habitat_baselines/common/trainer_registry.py
@@ -1,0 +1,27 @@
+from typing import Optional
+
+from habitat.core.registry import Registry
+
+
+class TrainRegistry(Registry):
+    @classmethod
+    def register_trainer(cls, to_register=None, *, name: Optional[str] = None):
+        r"""Register a RL training algorithm to registry with key 'name'
+
+        Args:
+            name: Key with which the trainer will be registered.
+                If None will use the name of the class
+
+        """
+        from habitat_baselines.common.base_model import BaseModel
+
+        return cls._register_impl(
+            "trainer", to_register, name, assert_type=BaseModel
+        )
+
+    @classmethod
+    def get_trainer(cls, name):
+        return cls._get_impl("trainer", name)
+
+
+train_registry = TrainRegistry()
