+++ habitat_baselines/config/default.py
@@ -22,12 +22,49 @@ _C.SEED = 100
 # BASELINE
 # -----------------------------------------------------------------------------
 _C.BASELINE = CN()
+_C.BASELINE.TRAINER_NAME = "ppo"
 # -----------------------------------------------------------------------------
 # REINFORCEMENT LEARNING (RL)
 # -----------------------------------------------------------------------------
 _C.BASELINE.RL = CN()
 _C.BASELINE.RL.SUCCESS_REWARD = 10.0
 _C.BASELINE.RL.SLACK_REWARD = -0.01
+# -----------------------------------------------------------------------------
+# PROXIMAL POLICY OPTIMIZATION (PPO)
+# -----------------------------------------------------------------------------
+_C.BASELINE.RL.PPO = CN()
+_C.BASELINE.RL.PPO.clip_param = 0.2
+_C.BASELINE.RL.PPO.ppo_epoch = 4
+_C.BASELINE.RL.PPO.num_mini_batch = 16
+_C.BASELINE.RL.PPO.value_loss_coef = 0.5
+_C.BASELINE.RL.PPO.entropy_coef = 0.01
+_C.BASELINE.RL.PPO.lr = 7e-4
+_C.BASELINE.RL.PPO.eps = 1e-5
+_C.BASELINE.RL.PPO.max_grad_norm = 0.5
+_C.BASELINE.RL.PPO.num_steps = 5
+_C.BASELINE.RL.PPO.hidden_size = 512
+_C.BASELINE.RL.PPO.num_processes = 16
+_C.BASELINE.RL.PPO.use_gae = True
+_C.BASELINE.RL.PPO.use_linear_lr_decay = False
+_C.BASELINE.RL.PPO.use_linear_clip_decay = False
+_C.BASELINE.RL.PPO.gamma = 0.99
+_C.BASELINE.RL.PPO.tau = 0.95
+_C.BASELINE.RL.PPO.log_file = "train.log"
+_C.BASELINE.RL.PPO.reward_window_size = 50
+_C.BASELINE.RL.PPO.log_interval = 1
+_C.BASELINE.RL.PPO.checkpoint_interval = 50
+_C.BASELINE.RL.PPO.checkpoint_folder = "data/checkpoints"
+_C.BASELINE.RL.PPO.sim_gpu_id = 0
+_C.BASELINE.RL.PPO.pth_gpu_id = 0
+_C.BASELINE.RL.PPO.num_updates = 10000
+_C.BASELINE.RL.PPO.sensors = "RGB_SENSOR,DEPTH_SENSOR"
+_C.BASELINE.RL.PPO.task_config = "configs/tasks/pointnav.yaml"
+_C.BASELINE.RL.PPO.tensorboard_dir = "tb"
+_C.BASELINE.RL.PPO.count_test_episodes = 2
+_C.BASELINE.RL.PPO.video_option = ["disk", "tensorboard"]
+_C.BASELINE.RL.PPO.video_dir = "video_Dir"
+_C.BASELINE.RL.PPO.tracking_model_dir = ""
+_C.BASELINE.RL.PPO.model_path = ""
 # -----------------------------------------------------------------------------
 # ORBSLAM2 BASELINE
 # -----------------------------------------------------------------------------
