+++ habitat_baselines/config/pointnav/ppo.yaml
@@ -1,0 +1,41 @@
+BASELINE:
+  TRAINER_NAME: "ppo"
+  RL:
+    PPO:
+      # general options
+      # TODO refactor these out of PPO
+      tensorboard_dir: "tb"
+      num_processes: 1
+      log_interval: 10
+      pth_gpu_id: 0
+      sim_gpu_id: 0
+      checkpoint_interval: 50
+      checkpoint_folder: "data/new_checkpoints"
+      task_config: "configs/tasks/pointnav.yaml"
+      num_updates: 10000
+      # eval specific:
+      count_test_episodes: 2
+      video_option: ["disk", "tensorboard"]
+      video_dir: "video_Dir"
+      model_path: "data/checkpoints/ckpt.123.pth"
+      tracking_model_dir: ""
+      # 1 of the above 2 options needs to be empty
+
+      # ppo params
+      clip_param: 0.1
+      ppo_epoch: 4
+      num_mini_batch: 1
+      value_loss_coef: 0.5
+      entropy_coef: 0.01
+      lr: 2.5e-4
+      eps: 1e-5
+      max_grad_norm: 0.5
+      num_steps: 128
+      hidden_size: 512
+      use_gae: True
+      gamma: 0.99
+      tau: 0.95
+      use_linear_clip_decay: True
+      use_linear_lr_decay: True
+      reward_window_size: 50
+      sensors: "RGB_SENSOR,DEPTH_SENSOR"
